+++ scripts/roop_swapping/upscaled_inswapper.py
@@ -11,11 +11,12 @@ from modules.face_restoration import FaceRestoration
 from modules.upscaler import UpscalerData
 from PIL import Image
 from scripts.roop_logging import logger
-from scripts.imgutils import cv2_to_pil, pil_to_cv2
+from scripts.roop_utils.imgutils import cv2_to_pil, pil_to_cv2
 from modules import processing
-from blendmodes.blend import blendLayers, BlendType
 from modules.shared import cmd_opts, opts, state
-from scripts.upscaling import UpscaleOptions, upscale_image
+from scripts.roop_postprocessing.postprocessing_options import PostProcessingOptions
+from scripts.roop_postprocessing import upscaling
+
 
 def get_upscaler() -> UpscalerData:
     for upscaler in shared.sd_upscalers:
@@ -35,13 +36,15 @@ class UpscaledINSwapper():
 
     def super_resolution(self,img, k = 2) :
         pil_img = cv2_to_pil(img)
-        upscaled =upscale_image(pil_img, UpscaleOptions(
+        options = PostProcessingOptions(
             upscaler_name=opts.data.get('roop_upscaled_swapper_upscaler', 'LDSR'),
             upscale_visibility=1,
             scale=k,
             face_restorer_name=opts.data.get('roop_upscaled_swapper_face_restorer', ""),
             codeformer_weight= opts.data.get('roop_upscaled_swapper_face_restorer_weight', 1),
-            restorer_visibility=opts.data.get('roop_upscaled_swapper_face_restorer_visibility', 1)))
+            restorer_visibility=opts.data.get('roop_upscaled_swapper_face_restorer_visibility', 1))
+        upscaled = upscaling.upscale_img(pil_img, options)
+        upscaled = upscaling.restore_face(upscaled, options)
         return pil_to_cv2(upscaled)
 
     def get(self, img, target_face, source_face, paste_back=True, upscale = True):
