--- scripts/roop_version.py
@@ -1,5 +1,0 @@
-version_flag = "v0.1.0"
-
-from scripts.roop_logging import logger
-
-logger.info(f"Roop {version_flag}")
